test refactor, new structure in place
