+++ brownie/cli/test.py
@@ -1,37 +1,11 @@
 #!/usr/bin/python3
 
-import ast
 from docopt import docopt
-from hashlib import sha1
-import importlib
-import json
-import os
-from pathlib import Path
-import re
-from requests.exceptions import ReadTimeout
-import sys
-import time
 
-from brownie.cli.utils import color
-from brownie.test import coverage
-from brownie.exceptions import ExpectedFailing
-import brownie.network as network
-from brownie.network.contract import Contract
-from brownie.network.history import TxHistory
-from brownie.project import build
-from brownie.types import FalseyDict
+from brownie.network import history
+from brownie.test.main import run_tests
 from brownie._config import ARGV, CONFIG
 
-COVERAGE_COLORS = [
-    (0.5, "bright red"),
-    (0.85, "bright yellow"),
-    (1, "bright green")
-]
-
-WARN = "{0[error]}WARNING{0}: ".format(color)
-ERROR = "{0[error]}ERROR{0}: ".format(color)
-
-history = TxHistory()
 
 __doc__ = """Usage: brownie test [<filename>] [<range>] [options]
 
@@ -59,399 +33,10 @@ def main():
         ARGV['always_transact'] = True
         history._revert_lock = True
 
-    # cleanup
-    project_path = Path(CONFIG['folders']['project'])
-    check_coverage_hashes(project_path)
-    recursive_unlink(project_path.joinpath('build/coverage'))
-
-    test_paths = get_test_paths(args['<filename>'])
-    coverage_files, test_data = get_test_data(test_paths)
-
-    if args['<range>']:
-        if len(test_paths) > 1:
-            sys.exit(ERROR+"Cannot specify a range when running multiple tests files.")
-        test_data = apply_range(test_data, args['<range>'])
-
-    if not test_data:
-        print("No tests to run.")
-    else:
-        run_test_modules(test_data, not args['<range>'])
-    if ARGV['coverage']:
-        display_report(coverage_files, not args['<range>'])
-    if ARGV['gas']:
-        display_gas_profile()
-
-
-def get_test_paths(path):
-    if not path:
-        path = ""
-    if path[:6] != "tests/":
-        path = "tests/" + path
-    path = Path(CONFIG['folders']['project']).joinpath(path)
-    if path.is_dir():
-        return [i for i in path.glob('**/*.py') if i.name[0] != "_" and "/_" not in str(i)]
-    if not path.suffix:
-        path = Path(str(path)+".py")
-    if not path.exists():
-        sys.exit(ERROR+"Cannot find {0[module]}tests/{1}{0}".format(color, path.name))
-    return [path]
-
-
-def check_coverage_hashes(base_path):
-    # remove coverage data where hashes have changed
-    coverage_path = Path(base_path).joinpath('build/coverage')
-    for coverage_json in list(coverage_path.glob('**/*.json')):
-        try:
-            dependents = json.load(coverage_json.open())['sha1']
-        except json.JSONDecodeError:
-            coverage_json.unlink()
-            continue
-        for path, hash_ in dependents.items():
-            path = Path(path)
-            if path.exists():
-                if path.suffix != ".json":
-                    if get_ast_hash(path) == hash_:
-                        continue
-                elif build.get(path.stem)['bytecodeSha1'] == hash_:
-                    continue
-            coverage_json.unlink()
-            break
-
-
-def recursive_unlink(base_path):
-    for path in [Path(i[0]) for i in list(os.walk(base_path))[:0:-1]]:
-        if not list(path.glob('*')):
-            path.rmdir()
-
-
-def get_test_data(test_paths):
-    coverage_files = []
-    test_data = []
-    project = Path(CONFIG['folders']['project'])
-    build_path = project.joinpath('build/coverage')
-    for path in test_paths:
-        path = path.relative_to(project)
-        coverage_json = build_path.joinpath(path.parent.relative_to('tests'))
-        coverage_json = coverage_json.joinpath(path.stem+".json")
-        coverage_files.append(coverage_json)
-        if coverage_json.exists():
-            coverage_eval = json.load(coverage_json.open())['coverage']
-            if ARGV['update'] and (coverage_eval or not ARGV['coverage']):
-                continue
-        else:
-            coverage_eval = {}
-            for p in list(coverage_json.parents)[::-1]:
-                p.mkdir(exist_ok=True)
-        module_name = str(path)[:-3].replace(os.sep, '.')
-        module = importlib.import_module(module_name)
-        test_names = re.findall(
-            r'\ndef[\s ]{1,}([^_]\w*)[\s ]*\([^)]*\)',
-            project.joinpath(path).open().read()
-        )
-        if not test_names:
-            print("\n{0}No test functions in {1[module]}{2}.py{1}".format(WARN, color, path))
-            continue
-        duplicates = set([i for i in test_names if test_names.count(i) > 1])
-        if duplicates:
-            raise ValueError("{} contains multiple test methods of the same name: {}".format(
-                path,
-                ", ".join(duplicates)
-            ))
-        if 'setup' in test_names:
-            args = _get_fn(module, 'setup')[-1]
-            if args['skip'] is True or (args['skip'] == "coverage" and ARGV['coverage']):
-                continue
-        test_data.append((module, coverage_json, coverage_eval, test_names))
-    return coverage_files, test_data
-
-
-def apply_range(test_data, idx):
-    test_names = list(test_data[0][-1])
-    if "0" in idx:
-        sys.exit(ERROR+"Range cannot include 0. First test is 1.")
-    idx = [i-1 if i > 0 else i for i in (int(x) for x in idx.split(':'))]
-    if len(idx) > 1:
-        if (min(idx+[0]), max(idx+[0])).count(0) != 1:
-            sys.exit(ERROR+"Range must be entirely positive or negative.")
-    idx = slice(*idx) if len(idx) > 1 else idx[0]
-    try:
-        if 'setup' in test_names:
-            test_names.remove('setup')
-            test_names = ['setup'] + ([test_names[idx]] if type(idx) is int else test_names[idx])
-        else:
-            test_names = [test_names[idx]] if type(idx) is int else test_names[idx]
-        return [list(test_data[0][:-1])+[test_names]]
-    except IndexError:
-        sys.exit(ERROR+"Invalid range. Must be an integer or slice (eg. 1:4)")
-
-
-def run_test_modules(test_data, save):
-    TestPrinter.grand_total = len(test_data)
-    count = sum([len([x for x in i[3] if x != "setup"]) for i in test_data])
-    print("Running {} tests across {} modules.".format(count, len(test_data)))
-    network.connect(ARGV['network'])
-    CONFIG._unlock()
-    CONFIG['active_network'].update(CONFIG['test'])
-    CONFIG._lock()
-    if not CONFIG['active_network']['broadcast_reverting_tx']:
-        print("{0[error]}WARNING{0}: Reverting transactions will NOT be broadcasted.".format(color))
-    traceback_info = []
-    start_time = time.time()
-    try:
-        for (module, coverage_json, coverage_eval, test_names) in test_data:
-            tb, cov, contracts = run_test(module, test_names)
-            if tb:
-                traceback_info += tb
-                if coverage_json.exists():
-                    coverage_json.unlink()
-                continue
-
-            if not save:
-                continue
-            if ARGV['coverage']:
-                coverage_eval = cov
-
-            contract_names = set(i._name for i in contracts)
-            contract_names |= set(x for i in contracts for x in i._build['dependencies'])
-
-            build_files = [Path('build/contracts/{}.json'.format(i)) for i in contract_names]
-
-            coverage_eval = {
-                'coverage': coverage_eval,
-                'sha1': dict((str(i), build.get(i.stem)['bytecodeSha1']) for i in build_files)
-            }
-            path = str(Path(module.__file__).relative_to(CONFIG['folders']['project']))
-            coverage_eval['sha1'][path] = get_ast_hash(module.__file__)
-            json.dump(
-                coverage_eval,
-                coverage_json.open('w'),
-                sort_keys=True,
-                indent=2,
-                default=sorted
-            )
-    except KeyboardInterrupt:
-        print("\n\nTest execution has been terminated by KeyboardInterrupt.")
-        sys.exit()
-    finally:
-        print("\nTotal runtime: {:.4f}s".format(time.time() - start_time))
-        if traceback_info:
-            print("{0}{1} test{2} failed.".format(
-                WARN,
-                len(traceback_info),
-                "s" if len(traceback_info) > 1 else ""
-            ))
-            for err in traceback_info:
-                print("\nTraceback for {0[0]}:\n{0[1]}".format(err))
-            sys.exit()
-    print("{0[success]}SUCCESS{0}: All tests passed.".format(color))
-
-
-def run_test(module, test_names):
-    network.rpc.reset()
-
-    if 'setup' in test_names:
-        test_names.remove('setup')
-        fn, default_args = _get_fn(module, 'setup')
-        p = TestPrinter(module.__file__, 0, len(test_names))
-        tb, coverage_eval = run_test_method(fn, default_args, {}, p)
-        if tb:
-            return tb, {}, set()
-    else:
-        p = TestPrinter(module.__file__, 1, len(test_names))
-        default_args = FalseyDict()
-        coverage_eval = {}
-    network.rpc.snapshot()
-    traceback_info = []
-    contracts = _get_contract_names()
-    for t in test_names:
-        history.clear()
-        network.rpc.revert()
-        fn, fn_args = _get_fn(module, t)
-        args = default_args.copy()
-        args.update(fn_args)
-        tb, coverage_eval = run_test_method(fn, args, coverage_eval, p)
-        contracts |= _get_contract_names()
-        traceback_info += tb
-        if tb and tb[0][2] == ReadTimeout:
-            print(WARN+"RPC crashed, terminating test")
-            network.rpc.kill(False)
-            network.rpc.launch(CONFIG['active_network']['test-rpc'])
-            break
-    if traceback_info and ARGV['coverage']:
-        coverage_eval = {}
-    p.finish()
-    return traceback_info, coverage_eval, contracts
-
-
-def run_test_method(fn, args, coverage_eval, p):
-    desc = fn.__doc__ or fn.__name__
-    if args['skip'] is True or (args['skip'] == "coverage" and ARGV['coverage']):
-        p.skip(desc)
-        return [], coverage_eval
-    p.start(desc)
-    try:
-        if ARGV['coverage'] and 'always_transact' in args:
-            ARGV['always_transact'] = args['always_transact']
-        fn()
-        if ARGV['coverage']:
-            ARGV['always_transact'] = True
-            coverage_eval = coverage.analyze(history.copy(), coverage_eval)
-        if args['pending']:
-            raise ExpectedFailing("Test was expected to fail")
-        p.stop()
-        return [], coverage_eval
-    except Exception as e:
-        p.stop(e, args['pending'])
-        if type(e) != ExpectedFailing and args['pending']:
-            return [], coverage_eval
-        path = Path(sys.modules[fn.__module__].__file__).relative_to(CONFIG['folders']['project'])
-        path = "{0[module]}{1}.{0[callable]}{2}{0}".format(color, str(path)[:-3], fn.__name__)
-        tb = color.format_tb(sys.exc_info(), sys.modules[fn.__module__].__file__)
-        return [(path, tb, type(e))], {}
-
-
-def display_report(coverage_files, save):
-    coverage_eval = coverage.merge_files(coverage_files)
-    report = {
-        'highlights': coverage.get_highlights(coverage_eval),
-        'coverage': coverage.get_totals(coverage_eval),
-        'sha1': {}  # TODO
-    }
-    print("\nCoverage analysis:")
-    for name in sorted(report['coverage']):
-        totals = report['coverage'][name]['totals']
-        pct = _pct(totals['statements'], totals['branches'])
-        c = color(next(i[1] for i in COVERAGE_COLORS if pct <= i[0]))
-        print("\n  contract: {0[contract]}{1}{0} - {2}{3:.1%}{0}".format(color, name, c, pct))
-        cov = report['coverage'][name]
-        for fn_name, count in cov['statements'].items():
-            branch = cov['branches'][fn_name] if fn_name in cov['branches'] else (0, 0, 0)
-            pct = _pct(count, branch)
-            print("    {0[contract_method]}{1}{0} - {2}{3:.1%}{0}".format(
-                color,
-                fn_name,
-                color(next(i[1] for i in COVERAGE_COLORS if pct <= i[0])),
-                pct
-            ))
-    if not save:
-        return
-    filename = "coverage-"+time.strftime('%d%m%y')+"{}.json"
-    path = Path(CONFIG['folders']['project']).joinpath('reports')
-    count = len(list(path.glob(filename.format('*'))))
-    path = path.joinpath(filename.format("-"+str(count) if count else ""))
-    json.dump(report, path.open('w'), sort_keys=True, indent=2, default=sorted)
-    print("\nCoverage report saved at {}".format(path.relative_to(CONFIG['folders']['project'])))
-
-
-def display_gas_profile():
-    print('\nGas Profile:')
-    gas = history.gas_profile
-    for i in sorted(gas):
-        print("{0} -  avg: {1[avg]:.0f}  low: {1[low]}  high: {1[high]}".format(i, gas[i]))
-
-
-def _pct(statement, branch):
-    pct = statement[0]/statement[1]
-    if branch[-1]:
-        pct = (pct + (branch[0]+branch[1])/(branch[2]*2)) / 2
-    return pct
-
-
-def _get_fn(module, name):
-    fn = getattr(module, name)
-    if not fn.__defaults__:
-        return fn, FalseyDict()
-    return fn, FalseyDict(zip(
-        fn.__code__.co_varnames[:len(fn.__defaults__)],
-        fn.__defaults__
-    ))
-
-
-class TestPrinter:
-
-    grand_count = 1
-    grand_total = 0
-
-    def __init__(self, path, count, total):
-        self.path = path
-        self.count = count
-        self.total = total
-        self.total_time = time.time()
-        print("\nRunning {0[module]}{1}{0} - {2} test{3} ({4}/{5})".format(
-            color,
-            path,
-            total,
-            "s" if total != 1 else "",
-            self.grand_count,
-            self.grand_total
-        ))
-
-    def skip(self, description):
-        self._print(
-            "{0} ({1[pending]}skipped{1[dull]})\n".format(description, color),
-            "\u229d",
-            "pending",
-            "dull"
-        )
-        self.count += 1
-
-    def start(self, description):
-        self.desc = description
-        self._print("{} ({}/{})...".format(description, self.count, self.total))
-        self.time = time.time()
-
-    def stop(self, err=None, expect=False):
-        if not err:
-            self._print("{} ({:.4f}s)  \n".format(self.desc, time.time() - self.time), "\u2713")
-        else:
-            err = type(err).__name__
-            color_str = 'success' if expect and err != "ExpectedFailing" else 'error'
-            symbol = '\u2717' if err in ("AssertionError", "VirtualMachineError") else '\u203C'
-            msg = "{} ({}{}{})\n".format(
-                self.desc,
-                color(color_str),
-                err,
-                color('dull')
-            )
-            self._print(msg, symbol, color_str, "dull")
-        self.count += 1
-
-    def finish(self):
-        print("Completed {0[module]}{1}{0} ({2:.4f}s)".format(
-            color,
-            self.path,
-            time.time() - self.total_time
-        ))
-        TestPrinter.grand_count += 1
-
-    def _print(self, msg, symbol=" ", symbol_color="success", main_color=None):
-        sys.stdout.write("\r {}{}{} {} - {}{}".format(
-            color(symbol_color),
-            symbol,
-            color(main_color),
-            self.count,
-            msg,
-            color
-        ))
-        sys.stdout.flush()
-
-
-def _get_contract_names():
-    return set(
-        i.contract_address for i in history if type(i.contract_address) is Contract and
-        not i.contract_address._build['sourcePath'].startswith('<string')
+    run_tests(
+        CONFIG['folders']['project'],
+        args['<filename>'],
+        ARGV['update'],
+        ARGV['coverage'],
+        ARGV['gas']
     )
-
-
-def get_ast_hash(script_path):
-    ast_list = [ast.parse(Path(script_path).open().read())]
-    for obj in [i for i in ast_list[0].body if type(i) in (ast.Import, ast.ImportFrom)]:
-        if type(obj) is ast.Import:
-            name = obj.names[0].name
-        else:
-            name = obj.module
-        origin = importlib.util.find_spec(name).origin
-        if CONFIG['folders']['project'] in origin:
-            ast_list.append(ast.parse(open(origin).read()))
-    dump = "\n".join(ast.dump(i) for i in ast_list)
-    return sha1(dump.encode()).hexdigest()
