+++ brownie/test/__init__.py
@@ -1,1 +1,3 @@
 #!/usr/bin/python3
+
+from .main import run_tests  # noqa F401
